Module Format.Date

include S

type t

val to_string : t -> string

val make : day:int -> month:int -> year:int -> t option

Create a date representation from a day, a month and a year. If the date is not a valid date string, None is returned